+++ grace-plugin-controllers/src/main/groovy/grails/artefact/Controller.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2025 the original author or authors.
+ * Copyright 2014-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -54,11 +54,11 @@ import org.grails.compiler.web.ControllerActionTransformer
 import org.grails.datastore.mapping.model.config.GormProperties
 import org.grails.plugins.web.api.MimeTypesApiSupport
 import org.grails.plugins.web.controllers.ControllerExceptionHandlerMetaData
-import org.grails.plugins.web.servlet.mvc.InvalidResponseHandler
-import org.grails.plugins.web.servlet.mvc.ValidResponseHandler
 import org.grails.web.servlet.mvc.GrailsWebRequest
+import org.grails.web.servlet.mvc.InvalidResponseHandler
 import org.grails.web.servlet.mvc.SynchronizerTokensHolder
 import org.grails.web.servlet.mvc.TokenResponseHandler
+import org.grails.web.servlet.mvc.ValidResponseHandler
 import org.grails.web.util.GrailsApplicationAttributes
 
 /**
